TSS-142: Unable to edit action plan user (#255)
* Tests for `SelectActionPlanOwner` fix.

* Extract `get_success_url` from `ActionPlanFormViewMixin` so it can be used independently of `get_form_kwargs`.

The original mixin's implementation of `get_form_kwargs` adds `barrier_id`. But the form used in `SelectActionPlanOwner`, `UserSearchForm`, isn't expecting that, and forms can't accept unknown keyword arguments. So to allow that view to reuse  `get_success_url` while not falling over because of the extra form kwarg, I've separated the two methods out.

* Shut Black up.

* Shut Black up on CircleCI.

Somehow, my local version of Black is changing this file thus, but the version of Black being run by the pre-commit hooks is changing it back. CircleCI isn't happy with _something_ so I'm going to see if it's this.

* Shut Black up.

* hotfix for max length on progress update (#257)

* hotfix for max length on progress update

* update pre-commit

* Revert "Shut Black up."

This reverts commit 32385a77afb32c64b266ad5a8c85431503361b10.

* Shut Black up. Maybe.

Co-authored-by: James Pacileo <608868+jamespacileo@users.noreply.github.com>