isort fixes
black and pep8

fixing tests

fixing tests and pep8 issues

refactoring tests to avoid form validation errors

Refactoring tests and deleting redundant code that was only strictly relevant to the 'add export source' page on the creation journey.

Fixed indentation in HTML

Editing 'export types' on the barrier dashboard now takes you to the new edit pages which include checkboxes for each export type, and a description of the exported goods.

Renamed export_type to export_types in the BarrierExportTypeForm to align with the DB

Deleted references to the old journey as no longer required.